File Downloading, with logging filtered

+++ email_monitor/email_reader.py
@@ -3,7 +3,13 @@ import email
 import os
 from datetime import datetime
 from dotenv import load_dotenv
+import logging
 
+logging.basicConfig(
+    filename="logs/processing.log",
+    level=logging.INFO,
+    format="%(asctime)s - %(message)s"
+)
 load_dotenv()
 
 EMAIL = os.getenv("EMAIL_ADDRESS")
@@ -17,13 +23,28 @@ def connect_email():
     mail.login(EMAIL, PASSWORD)
     return mail
 
-def fetch_unread_emails(mail):
+from datetime import datetime, timedelta
+
+from datetime import datetime, timedelta
+
+def fetch_recent_emails(mail):
+
     mail.select("inbox")
-    status, messages = mail.search(None, '(UNSEEN)')
+
+    since_date = (datetime.now() - timedelta(days=2)).strftime("%d-%b-%Y")
+
+    print("Searching emails since:", since_date)
+
+    status, messages = mail.search(None, 'SINCE', since_date)
+
     email_ids = messages[0].split()
+
+    print("Emails found:", len(email_ids))
+
     return email_ids
 
-def save_attachment(part, filename):
+def save_attachment(file_data, filename):
+
     today = datetime.now().strftime("%Y-%m")
     folder = f"bills/{today}"
 
@@ -31,14 +52,18 @@ def save_attachment(part, filename):
 
     filepath = os.path.join(folder, filename)
 
+    if os.path.exists(filepath):
+        print(f"File already exists, skipping: {filename}")
+        return
+
     with open(filepath, "wb") as f:
-        f.write(part.get_payload(decode=True))
+        f.write(file_data)
 
     print(f"Saved attachment: {filepath}")
 
 def process_emails():
     mail = connect_email()
-    email_ids = fetch_unread_emails(mail)
+    email_ids = fetch_recent_emails(mail)
 
     for e_id in email_ids:
         status, msg_data = mail.fetch(e_id, "(RFC822)")
@@ -47,14 +72,21 @@ def process_emails():
             if isinstance(response_part, tuple):
 
                 msg = email.message_from_bytes(response_part[1])
-
                 for part in msg.walk():
 
-                    if part.get_content_disposition() == "attachment":
-                        filename = part.get_filename()
+                    filename = part.get_filename()
+                    allowed_extensions = [".pdf", ".png", ".jpg", ".jpeg"]
 
-                        if filename:
-                            save_attachment(part, filename)
+                    if filename and any(filename.lower().endswith(ext) for ext in allowed_extensions):
+
+                        file_data = part.get_payload(decode=True)
+                        file_size = len(file_data)
+
+                        if file_size < 5000:
+                            print("Skipping small file:", filename)
+                            continue
+
+                        save_attachment(file_data, filename)
 
 if __name__ == "__main__":
     process_emails()
